Adds ContentLoader component to Library
