+++ src/components/index.js
@@ -21,14 +21,15 @@ export { default as Button } from './button';
 export { default as CheckBoxControl } from './checkbox-control';
 export { default as ClipPathControl } from './clip-path-control';
 export { default as ColorControl } from './color-control';
+export { default as ContentLoader } from './content-loader';
 export { default as CustomCss } from './custom-css-control';
 export { default as CustomLabel } from './custom-label';
 export { default as DefaultStylesControl } from './default-styles-control';
 export { default as DialogBox } from './dialog-box';
-export { default as DynamicContent } from './dynamic-content';
 export { default as DisplayControl } from './display-control';
 export { default as DividerControl } from './divider-control';
 export { default as Dropdown } from './dropdown';
+export { default as DynamicContent } from './dynamic-content';
 export { default as FlexAlignControl } from './flex-settings-control/flex-align-control';
 export { default as FlexGapControl } from './flex-settings-control/flex-gap-control';
 export { default as FlexSettingsControl } from './flex-settings-control';
@@ -55,6 +56,7 @@ export { default as OpacityControl } from './opacity-control';
 export { default as OverflowControl } from './overflow-control';
 export { default as PaddingControl } from './padding-control';
 export { default as Placeholder } from './placeholder';
+export { default as Popover } from './popover';
 export { default as PositionControl } from './position-control';
 export { default as RawHTML } from './raw-html';
 export { default as RelationControl } from './relation-control';
@@ -80,4 +82,3 @@ export { default as TransformControl } from './transform-control';
 export { default as TransitionControl } from './transition-control';
 export { default as TypographyControl } from './typography-control';
 export { default as ZIndexControl } from './zindex-control';
-export { default as Popover } from './popover';
