+++ src/editor/library/container.js
@@ -17,6 +17,8 @@ import MasonryItem from './MasonryItem';
 import masonryGenerator from './masonryGenerator';
 import useInterval from '../../extensions/dom/useInterval';
 import InfiniteHits from './InfiniteHits';
+import onRequestInsertPattern from './utils/onRequestInsertPattern';
+import { ContentLoader } from '../../components';
 
 /**
  * External dependencies
@@ -40,7 +42,6 @@ import DOMPurify from 'dompurify';
  * Icons
  */
 import { arrowIcon } from '../../icons';
-import onRequestInsertPattern from './utils/onRequestInsertPattern';
 
 // hack to fix issue #3930: top level tags resetting when we choose a second-level tag
 const removeMenuBugFix = () => {
@@ -86,7 +87,7 @@ const LoadingContent = () => (
 			width: '100%',
 		}}
 	>
-		Loading!
+		<ContentLoader />
 	</div>
 );
